new files

+++ app/Http/Controllers/UserController.php
@@ -23,35 +23,42 @@ class UserController extends Controller
     public function updateProfile(Request $request)
     {
         $user = Auth::user();
-
+    
         $request->validate([
             'profile_image' => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
             'telegram' => 'required|string|unique:users,telegram,' . $user->id,
         ]);
+    
         try {
             if ($request->hasFile('profile_image')) {
+                // Caminho para a pasta profile_images dentro de public_html
+                $profileImagesPath = base_path('../public_html/profile_images');
+    
                 // Delete old profile image if exists
-                if ($user->profile_image && Storage::exists('public/profile_img/' . $user->profile_image)) {
-                    Storage::delete('public/profile_img/' . $user->profile_image);
+                if ($user->profile_image && file_exists($profileImagesPath . '/' . $user->profile_image)) {
+                    unlink($profileImagesPath . '/' . $user->profile_image);
                 }
-
-                // Store new profile image
+    
+                // Armazenar nova imagem de perfil
                 $profileImageName = $user->id . '_' . time() . '.' . $request->profile_image->getClientOriginalExtension();
-                $request->profile_image->storeAs('public/profile_img', $profileImageName);
-
-                // Update user profile image
+                $request->profile_image->move($profileImagesPath, $profileImageName);
+    
+                // Atualizar imagem de perfil do usuário
                 $user->profile_image = $profileImageName;
             }
-
-            // Update telegram username
+    
+            // Atualizar o nome de usuário do Telegram
             $user->telegram = $request->telegram;
             $user->save();
-
+    
             return redirect()->route('profile')->with('success', 'Profile updated successfully.');
         } catch (\Exception $e) {
             return redirect()->back()->with('error', 'An error occurred when updating the profile. Please try again.');
         }
     }
+
+
+
 
     public function accountSettings(User $user)
     {
@@ -108,7 +115,7 @@ class UserController extends Controller
         // Validação dos campos
         $fields = $request->validate([
             'name' => 'required',
-            'email' => 'required|email|ends_with:@elpato.com',
+            'email' => 'required|email|ends_with:@elpato.xyz',
             'password' => [
                 'required',
                 'min:8',
@@ -117,12 +124,15 @@ class UserController extends Controller
             'email_verified_at' => 'required|date',
             'type' => 'required',
             'telegram' => 'required',
+            'g-recaptcha-response' => 'required|captcha', // Validação do reCAPTCHA
         ], [
-            'email.ends_with' => 'The email must be a valid @elpato.com email address.',
+            'email.ends_with' => 'The email must be a valid @elpato.xyz email address.',
             'password.min' => 'The password must be at least 8 characters long.',
-            'password.regex' => 'The password must contain at least 1 special character.'
+            'password.regex' => 'The password must contain at least 1 special character.',
+            'g-recaptcha-response.required' => 'Please confirm that you are not a robot.',
+            'g-recaptcha-response.captcha' => 'reCAPTCHA verification failed. Please try again.',
         ]);
-
+    
         try {
             // Criação do novo usuário
             $user = new User();
@@ -133,10 +143,10 @@ class UserController extends Controller
             $user->type = $fields['type'];
             $user->telegram = $fields['telegram'];
             $user->blocked = 0; // Define o valor padrão para 'blocked'
-
+    
             // Salvando o usuário no banco de dados
             $user->save();
-
+    
             return redirect()->route('user.all')->with('success', 'User created successfully!');
         } catch (\Exception $e) {
             return redirect()->back()->with('error', 'An error occurred while creating the new user. Please try again.');
@@ -145,10 +155,11 @@ class UserController extends Controller
 
 
 
+
     public function setDefaultPassword($id)
     {
         try {
-            $defaultPassword = '!elpato12345';
+            $defaultPassword = '3fFh@p8J0QJ74#';
             $user = User::findOrFail($id);
             $user->password = bcrypt($defaultPassword);
             $user->save();
@@ -206,46 +217,59 @@ class UserController extends Controller
      * Update the specified resource in storage.
      */
     public function update(Request $request, $id)
-    {
-        // Validação dos campos
-        $fields = $request->validate([
-            'name' => 'required',
-            'email' => 'required|email|ends_with:@elpato.com',
-            'email_verified_at' => 'required|date',
-            'type' => 'required',
-            'telegram' => 'required',
-            'blocked' => 'required|boolean',
-        ], [
-            'email.ends_with' => 'The email must be a valid @elpato.com email address.',
-        ]);
+{
+    // Validação dos campos
+    $fields = $request->validate([
+        'name' => 'required',
+        'email' => 'required|email|ends_with:@elpato.xyz',
+        'email_verified_at' => 'required|date',
+        'type' => 'required',
+        'telegram' => 'required',
+        'blocked' => 'required|boolean',
+    ], [
+        'email.ends_with' => 'The email must be a valid @elpato.xyz email address.',
+    ]);
 
-        try {
-            $user = User::findOrFail($id);
+    try {
+        $user = User::findOrFail($id);
 
-            // Impede o usuário de alterar seu próprio estado de 'blocked'
-            if (auth()->user()->id == $id) {
-                return redirect()->back()->with('error', 'You cannot change your own blocked status.');
-            }
-
-            $user->name = $fields['name'];
-            $user->email = $fields['email'];
-            $user->email_verified_at = $fields['email_verified_at'];
-            $user->type = $fields['type'];
-            $user->telegram = $fields['telegram'];
-            $user->blocked = $fields['blocked'] === '1' ? true : false;
-
-            $user->save();
-
-            return redirect()->route('user.all')->with('success', 'User updated successfully!');
-        } catch (\Exception $e) {
-            return redirect()->back()->with('error', 'An error occurred while updating the user. Please try again.');
+        // Impede o usuário de alterar seu próprio estado de 'blocked'
+        if (auth()->user()->id == $id) {
+            return redirect()->back()->with('error', 'You cannot change your own blocked status.');
         }
+
+        // Impede um admin de mudar a role e o estado de bloqueio de outro admin
+        if (auth()->user()->type === 'admin' && $user->type === 'admin') {
+            if ($fields['type'] !== $user->type) {
+                return redirect()->back()->with('error', 'You cannot change the role of another admin.');
+            }
+            if ($fields['blocked'] != $user->blocked) {
+                return redirect()->back()->with('error', 'You cannot change the blocked status of another admin.');
+            }
+        }
+
+        $user->name = $fields['name'];
+        $user->email = $fields['email'];
+        $user->email_verified_at = $fields['email_verified_at'];
+        $user->type = $fields['type']; // Isso só será permitido se não for um admin alterando outro admin
+        $user->telegram = $fields['telegram'];
+        $user->blocked = $fields['blocked'] === '1' ? true : false;
+
+        $user->save();
+
+        return redirect()->route('user.all')->with('success', 'User updated successfully!');
+    } catch (\Exception $e) {
+        return redirect()->back()->with('error', 'An error occurred while updating the user. Please try again.');
     }
+}
 
 
     /**
      * Remove the specified resource from storage.
      */
+    /**
+ * Remove the specified resource from storage.
+ */
     public function destroy(User $user)
     {
         try {
@@ -253,16 +277,24 @@ class UserController extends Controller
             if (auth()->user()->id === $user->id) {
                 return redirect()->route('user.all')->with('error', 'You cannot delete your own account.');
             }
-
+    
+            // Prevent deletion of other administrators
+            if ($user->type === 'admin') {
+                return redirect()->route('user.all')->with('error', 'You cannot delete other administrators.');
+            }
+    
             // Delete profile image if exists
             if ($user->profile_image && Storage::exists('public/profile_img/' . $user->profile_image)) {
                 Storage::delete('public/profile_img/' . $user->profile_image);
             }
+    
+            // Delete the user
             $user->delete();
-
+    
             return redirect()->route('user.all')->with('success', 'User deleted successfully!');
         } catch (\Exception $e) {
             return redirect()->back()->with('error', 'An error occurred while deleting the user. Please try again.');
         }
     }
+
 }
